Eucler Circuit in an Undirectd Graph
Eulerian Path is a path in a graph that visits every edge exactly once. Eulerian Circuit is an Eulerian Path that starts and ends on the same vertex. Given the number of vertices v and adjacency list adj denoting the graph. Find that there exists the Euler circuit or not. Return 1 if there exist  alteast one eulerian path else 0.
Example 1:

Input: 
v = 4 
edges[] = {{0, 1}, 
           {0, 2}, 
           {1, 3}, 
           {2, 3}}

Output: 
1
Explanation: corresponding adjacency list will be {{1, 2},{0, 3},{0, 3},{1, 2}}
One of the Eularian circuit 
starting from vertex 0 is as follows:
0->1->3->2->0